WebSocket: Wait graciously 5 seconds for server close response before shutdown
There might be out there a bad server which doesn't reply to a close
request, keeping the connection alive.

If the client explicity wants to close a WebSocket connection by calling
the close() method, there's no point in keeping the connection alive.

So, let's wait 5 seconds for a graceful response from server. If it doesn't
arrive, just close the connection anyway.

Closes #2788.