Prevent NPE in Reponse#close() when body is null
The `Response#close()` method threw a `NullPointerException` if it was called and
the response body was `null`.

Since Response is `Closable` (and transitively `AutoClosable`), it should be possible
to use a `Response `object in a try-with-resources block without thinking about whether
the `#close()` method would throw an exception in some cases.

Example:

    try (Response response = client.newCall(call).execute()) {
      // do something with response
    }