We should always have our callbacks called
A common user error with OkHttp is to write an interceptor which might
throw an IllegalArgumentException or similar. This works fine when using
the 'execute()' interface, but with 'enqueue()' leads to serious
problems; callbacks might never be fired.

This leads to very subtle bugs. For example, the Retrofit Java 8 call
adapter returns a CompletableFuture, which might never be completed.

Here, we wrap every throwable with an IOException to ensure that the
callback fires.