Close server UNIX domain sockets gracefully
On OS X, if close() is called while accept() is blocking, errno is set to ECONNABORTED. UnixServerSocketChannel.accept() propagates this as an IOException. This is incompatible with java.net.ServerSocket, which throws a SocketException in the same scenario. MockWebServer expects a SocketException in case of a clean shutdown and incorrectly reports an error if an IOException is thrown instead.
To prevent this, we try to detect this scenario and throw a SocketException.

This fixes https://github.com/square/okhttp/issues/4106.