Don't fail the build if dependencies aren't available
This causes us to fail if our own snapshots are missing. I wasn't able to
get the normal Maven options to do this, including  -DexcludeGroupIds=com.squareup.okhttp3
and -DexcludeTransitive=true which I expected to work but don't.

My expectation is that if dependencies aren't available, the build will soon
fail anyway.