Move tests from okhttp-tests/src/test/java to okhttp/src/test/java
The previous module structure was necessary because Maven wouldn't allow
mockwebserver to depend on okhttp/src and for okhttp/test to depend on
mockwebserver. With Gradle this constraint is lifted and we can fold
everything into a single module.

I'm in a rush to do this now because it's necessary for Kotlin 'internal'
tests to be in the same module.