Correctly calculate HTTP2 response body size
Currently response body size is determined by Content-Length header value, i.e. we use the header value if it exists, otherwise set -1.

This doesn't work for some HTTP responses. For example, `response.body().contentLength()` should return 0 for a response of HEAD request instead of using the Content-Length value. Also it should return 0 for a 204 responses instead of providing -1 when Content-Length header is not present.