Switch to Gradle's Maven Publishing Plugin
This removes a bunch of low-value stuff from the pom.xml files:

 - name
 - description
 - licence distribution ('repo')
 - developers clause
 - test dependencies

I don't think any of this will be missed, and it shrinks the pom.xml
file to the minimal set of useful stuff.

This also causes us to publish a gradle .module file. This is the
motivation for this change. It'll allow us to ship a Gradle platform,
which is a more capable than a Maven BOM.