Don't do infinite retries on server-canceled calls
We originally had a bug where Call.cancel() would incorrectly
cause HTTP/2 connections to be closed.

The fix was to not consider HTTP/2 connections to be degraded
when a CANCEL error is received:
https://github.com/square/okhttp/pull/4052

This closes the loop to make sure that we only treat CANCEL as
expected when the call was actually canceled.

Closes: https://github.com/square/okhttp/issues/5726