Fix errorprone PreferJavaTimeOverload in tests
When compiling OkHttp, we have the warnings regarding PreferJavaTimeOverload.

For example:
```
okhttp/okhttp/src/test/java/okhttp3/EventListenerTest.java:193: warning: [PreferJavaTimeOverload] Prefer using java.time-based APIs when available. Note that this checker does not and cannot guarantee that the overloads have equivalent semantics, but that is generally the case with overloaded methods.
        .readTimeout(250, TimeUnit.MILLISECONDS)
                    ^
    (see https://errorprone.info/bugpattern/PreferJavaTimeOverload)
```

To fix this, `java.time.Duration` API is used instead of `java.util.concurrent.TimeUnit` in tests.