Rip out WebSocketWriter's message splitting
An old version of OkHttp web sockets supported streaming messages,
so a writer could transmit partial bytes of a message without
completing it.

When we promoted Web Sockets as a core feature we dropped support
for streaming a single message. But we still had the mechanism
that implemented it.

This PR removes that mechanism, which should make it easier to
implement compression.