Tidy up the code that does connection releasing in RealCall
Recent refactorings have made it clear that there's three things
we need to be done before we can release the connection:
 - the request body
 - the response body
 - the series of exchanges

This changes the code to invert these booleans. In a follow-up I'd
like to add some defensive code around duplex calls where the request
and response body could race when they close.