Test interaction between pins and non-CA certificates (#6586)
Our path building is very naive and it's possible to cause OkHttp to
create chains that are different from the chains previously verified
by the SSL engine.

Our preference is to use the platform's chain where possible. This includes
Android API 17+, which has an API that returns the actual chain that
was verified, not the (potentially larger) set of certificates served
by the server.

We should further lock down our chain processing. I'm reluctant to do
full graph building because it's a lot of ugly code and potentially
expensive. For now, a test that attempts to detect gaps between what
the TLS layer lets through and what our chain cleaner computes.