Fixed leading spaces for code blocks in recipes.md
Markdown coverts text with four leading spaces into a code block. Fenced code blocks in this file had four leading spaces in addition to the being wrapped with ```, because of this the code blocks weren't being rendered properly in the markdown preview.

Removing the two out of the four leading spaces fixed the issue.