Drop the Content-Length header from multipart parts
This causes our users grief:
https://github.com/square/okhttp/issues/2604
https://github.com/square/okhttp/issues/2138

I've resisted removing it because it seemed like it'd
break compatibility in a significant way. But with OkHttp 5
it's a good time to break compatibility.

As far as I can tell the behavior of including Content-Length
in multipart bodies is not performed by web browsers or other
HTTP clients.