Make jvm functions compatible with kotlin 1.5 runtime
Based on the discovery in https://github.com/square/okhttp/pull/7329#issuecomment-1159192273
this commit tries a different approach to make okhttp compatible
with the kotlin 1.5 runtime. Specifically, it converts the DurationUnit
instance to a java.util.concurrent.TimeUnit instance and uses the
existing codepath to handle it. The conversion must be done without
using `DurationUnit.toTimeUnit()` since that is again a function
that would not exist in the 1.5 kotlin runtime environment.