Create MockResponseBody (#7593)
* Create MockResponseBody

We'd previously only ever used Buffer to model the type
of the MockResponse body. This has proven inadequate, particularly
for features like duplex responses.

Most of the complexity in this PR is changing how throttling
and disconnect-during-body work when we don't necessarily have
a Buffer for the response body.

This is a first step, more capability to follow.

* Update API

* We flush fewer times now

* Attempt to make a test less flaky

* Try harder to make the test less flaky