Implement a compact IDNA mapping table
This is optimized for fast lookups and small code size. It also
avoids using a resource file, as that's not an option on Kotlin/JS.

The data is encoded with 3 ASCII strings, as described in the new
IdnaMappingTable class docs. I've got tests that confirm the compact
data is consistent with the simple implementation for all code points.