Fix a concurrency problem with web socket cleanup
We were seeing crashes in Deflater.deflate with a root cause
of closing the deflater while it was in use.

The underlying problem is that web socket streams were closed
when the web socket was canceled, but not by the thread that
owned those streams.

This moves stream cleanup to always run on the owning thread:
 - writer cleanup happens on the task queue
 - reader cleanup happens after loopReader

This also introduces a new function, Streams.cancel() so a
failed writer can break an active reader, and vise-versa.
Previously we were using close() to do this job.

There's a lot of test changes but it's mostly in adding
the TaskFaker facet to RealWebSocketTest. The test now
runs with a fake clock, with the deterministic benefits
that brings.

See also: https://github.com/square/okhttp/issues/6719