Allow AddressPolicy to have a max calls per connection
This adds the ability for the Address Policy to also contain a maximum
number of connections per Stream.

In some environments it seems that OkHttp's HTTP/2 gets worse the more
streams that are packed onto a single connection. By adding in the
ability for users to alter their client's AddressPolicy to set a maximum
number of connections per stream, they can control how much (if any)
packing they want OkHttp to perform.