Use SHOUTY_CASE for static constants (#8802)
OkHttp has lots of constants defined from before it was a
Kotlin library, and I'd rather the library be self-consistent
than consistent with Kotlin idioms.

As a bonus we get to remain consistent with Java idioms.
I'd love a Kotlin mechanism to use @JvmField and @JvmName
at the same time, but I can't figure out how.

Co-authored-by: Jesse Wilson <jwilson@squareup.com>