TrailersSource, a new public API for HTTP trailers (#8829)
* TrailersSource, a new public API for HTTP trailers

This doesn't change much.

The new interface replaces a Kotlin lambda that had
similar behavior, though that lambda wasn't specified
as thoroughly.

We now permit calls to Response.trailers() before the
entire response body is consumed. But when such calls are
made the response body is discarded first. This could be
a bit of a surprise to users, especially users who do an
additional layer of buffering over their ResponseBody
because it could yield latent failures.

But I think it's better behavior overall, especially since
it's possible data binding layers might not consume the
entire response body in all cases. For example, a JSON
library might not read to the very end of the response
if it has received a terminating '}' byte.

The riskiest part of this change is when ResponseBodySource
self-reports as complete. Previously it would self-report
as complete when it returned as many bytes as promised
in the Content-Length. With this change it will now only
report itself complete when it receives a definitive EOF
on the stream, signaled by a -1 byte count on a read. This
is because only when the EOF is received can we be sure
that the trailers are received, and we must not unhook
the Exchange from the Call until that happens. The previous
behavior could make Call.cancel() no-op even if we were
blocked waiting on trailers.

* apiDump

* Attempt to get more tests passing

* Don't wait for stream EOF to release the stream

One of our tests caught a situation where we were waiting
for the caller to read the EOF when we didn't need to.

---------

Co-authored-by: Jesse Wilson <jwilson@squareup.com>