Introduce events for the dispatcher queue
So event listeners can see how long calls are waiting for
resources to become available.

This is the API only. I'll do the implementation in
follow-up.

Also note that I'm going to try my best to implement
this event pair to only fire for calls that actually
spend time in the queue. Calls that are immediately
promoted out of the queue should not get events.