Use a single synchronized block in Dispatcher (#9110)
Previously we had synchronized blocks preceding calls
to promoteAndExecute, plus the synchronized blocks in
that function itself.

This is intended to make it easier to publishg the right
events for dispatcherQueueStart and dispatcherQueueEnd
when an enqueued call can skip the queue.

Also note that this should fix some corner-cases around
unnecessary calls to idleCallback when the executor is
already shut down.

Co-authored-by: Jesse Wilson <jwilson@squareup.com>