New datastructure to hold Tags (#9163)
* New datastructure to hold Tags

We have been using MutableMap and Map. These have worked
fine.

But we'd like to soon support mutating tags on an existing
Request object, which I'd like to do without a ton of
collections or concurrency overhead. Instead we can do some
lock-free stuff, which needs a simple data structure as a
starting point.

* Fill in some type parameters